splits views.py into modules, refactor import in app.py
